requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
azure-core==1.29.4
azure-identity==1.14.1
azure-storage-blob==12.18.3
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
cryptography==41.0.4
cymem==2.0.8
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
docx==0.2.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
Flask==2.3.3
frozenlist==1.4.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
langchain==0.0.316
langcodes==3.3.0
langsmith==0.0.44
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
msal==1.24.1
msal-extensions==1.0.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
openpyxl==3.1.2
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-aws-lambda==0.41b0
opentelemetry-instrumentation-dbapi==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-grpc==0.41b0
opentelemetry-instrumentation-jinja2==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-sqlite3==0.41b0
opentelemetry-instrumentation-system-metrics==0.41b0
opentelemetry-instrumentation-tornado==0.41b0
opentelemetry-instrumentation-tortoiseorm==0.41b0
opentelemetry-instrumentation-urllib==0.41b0
opentelemetry-instrumentation-urllib3==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pypdf==3.16.4
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.0.1
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacey==0.1.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.22
srsly==2.4.8
stack-data==0.6.3
tenacity==8.2.3
thinc==8.2.1
tiktoken==0.5.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
validators==0.22.0
wasabi==1.1.2
waybackpy==3.0.6
wcwidth==0.2.8
weasel==0.3.3
weaviate-client==3.24.2
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.15.0
XlsxWriter==3.1.8
yarl==1.9.2
zipp==3.17.0
